Rename query and adjust namespace
