+++ src/SFA.DAS.CourseDelivery.Api/Controllers/CoursesController.cs
@@ -7,8 +7,9 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using SFA.DAS.CourseDelivery.Api.ApiRequests;
 using SFA.DAS.CourseDelivery.Api.ApiResponses;
-using SFA.DAS.CourseDelivery.Application.Provider.Queries.Provider;
+using SFA.DAS.CourseDelivery.Application.Provider.Queries.ProviderByCourse;
 using SFA.DAS.CourseDelivery.Application.Provider.Queries.ProvidersByCourse;
+using GetCourseProviderResponse = SFA.DAS.CourseDelivery.Api.ApiResponses.GetCourseProviderResponse;
 using GetProviderResponse = SFA.DAS.CourseDelivery.Api.ApiResponses.GetProviderResponse;
 
 namespace SFA.DAS.CourseDelivery.Api.Controllers
@@ -63,11 +64,11 @@ namespace SFA.DAS.CourseDelivery.Api.Controllers
         
         [HttpGet]
         [Route("{id}/providers/{ukprn}")]
-        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GetProviderResponse))]
-        [ProducesResponseType((int)HttpStatusCode.NotFound, Type = typeof(GetProviderResponse))]
+        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GetCourseProviderResponse))]
+        [ProducesResponseType((int)HttpStatusCode.NotFound, Type = typeof(GetCourseProviderResponse))]
         public async Task<IActionResult> GetProviderByUkprn(int id, int ukprn)
         {
-            var queryResult = await _mediator.Send(new GetProviderQuery {Ukprn = ukprn, StandardId = id});
+            var queryResult = await _mediator.Send(new GetCourseProviderQuery {Ukprn = ukprn, StandardId = id});
 
             if (queryResult.ProviderStandardContact == null)
             {
