--- src/SFA.DAS.CourseDelivery.Application/Provider/Queries/Provider/GetProviderQuery.cs
@@ -1,10 +1,0 @@
-using MediatR;
-
-namespace SFA.DAS.CourseDelivery.Application.Provider.Queries.Provider
-{
-    public class GetProviderQuery : IRequest<GetProviderResponse>
-    {
-        public int Ukprn { get ; set ; }
-        public int StandardId { get ; set ; }
-    }
-}
