+++ src/SFA.DAS.CourseDelivery.Application/Provider/Queries/ProviderByCourse/GetCourseProviderQuery.cs
@@ -1,0 +1,10 @@
+using MediatR;
+
+namespace SFA.DAS.CourseDelivery.Application.Provider.Queries.ProviderByCourse
+{
+    public class GetCourseProviderQuery : IRequest<GetCourseProviderResponse>
+    {
+        public int Ukprn { get ; set ; }
+        public int StandardId { get ; set ; }
+    }
+}
